Move interface file to types dir and add _layout file in app dir
